+++ SofaKernel/framework/framework_test/CMakeLists.txt
@@ -29,6 +29,9 @@ set(SOURCE_FILES
     helper/system/atomic_test.cpp
     helper/logging/logging_test.cpp
     testing/TestMessageHandler_test.cpp
+    simulation/TaskSchedulerTestTasks.h
+    simulation/TaskSchedulerTests.cpp
+    simulation/TaskSchedulerTestTasks.cpp
     main.cpp
 )
 
@@ -41,7 +44,7 @@ include_directories(${gtest_SOURCE_DIR}/include)
 add_definitions("-DFRAMEWORK_TEST_RESOURCES_DIR=\"${CMAKE_CURRENT_SOURCE_DIR}/resources\"")
 
 add_executable(${PROJECT_NAME} ${SOURCE_FILES})
-target_link_libraries(${PROJECT_NAME} gtest SofaHelper SofaCore SofaDefaultType)
+target_link_libraries(${PROJECT_NAME} gtest SofaHelper SofaCore SofaDefaultType SofaSimulationCore)
 #add_dependencies(${PROJECT_NAME} PluginA PluginB PluginC PluginD PluginE PluginF)
 
 add_test(NAME ${PROJECT_NAME} COMMAND ${PROJECT_NAME})
