+++ SofaKernel/framework/framework_test/simulation/TaskSchedulerTestTasks.h
@@ -1,4 +1,4 @@
-#include <MultiThreading/src/Task.h>
+#include <sofa/simulation/Task.h>
 
 
 namespace sofa
