+++ applications/plugins/MultiThreading/test/CMakeLists.txt
@@ -11,11 +11,10 @@ set(SOURCE_FILES
 )
 
 find_package(SofaTest REQUIRED)
-find_package(MultiThreading REQUIRED)
 
 add_executable(${PROJECT_NAME} ${SOURCE_FILES} ${HEADER_FILES})
 #add_test(NAME ${PROJECT_NAME} COMMAND ${PROJECT_NAME})
-target_link_libraries(${PROJECT_NAME} SofaTest SofaGTestMain MultiThreading)
-target_include_directories(${PROJECT_NAME} PRIVATE ${MultiThreading_INCLUDE_DIRS})
+target_link_libraries(${PROJECT_NAME} SofaTest SofaGTestMain SofaSimulationCore)
+#target_include_directories(${PROJECT_NAME} PRIVATE ${SofaSimulationCore_INCLUDE_DIRS})
 
 add_test(NAME ${PROJECT_NAME} COMMAND ${PROJECT_NAME})
